+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/kafkastreams/internal/utils/ProcessorContextLogging.scala
@@ -9,11 +9,6 @@ trait ProcessorContextLogging {
 
   private val _logger = Logger(getClass)
 
-  @deprecated("Use error, warn, info, debug, or trace methods directly")
-  protected def logger: Logger = {
-    _logger
-  }
-
   protected def processorContext: ProcessorContext
 
   final protected[this] def error(message: => Any): Unit = {