+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/kafkastreams/internal/utils/sampling/ReservoirSamplingTransformer.scala
@@ -66,9 +66,7 @@ class ReservoirSamplingTransformer[
     sampleStore
       .deleteRange(
         IndexedSampleKey.rangeStart(key),
-        IndexedSampleKey.rangeEnd(key),
-        maxDeletes = sampleSize
-      )
+        IndexedSampleKey.rangeEnd(key))
 
     numExpiredCounter.incr()
   }